Move remainder of WindowBackend into Window
